Updating youtube-dlp

+++ src/main/java/EncodingMagic/URLPegGenerator.java
@@ -101,7 +101,7 @@ public class URLPegGenerator implements PegGenerator {
     }
 
     @Override
-    public void passMetaData(double source_fps, double source_duration) {
+    public void passMetaData(float source_fps, float source_duration) {
         this.sourceFps = source_fps;
         this.sourceTotalFrames = (int) (source_duration * sourceFps);
     }
